refactor: 운동 게스트 초대 API에서 중첩 record 패턴 적용
