--- src/main/java/umc/cockple/demo/domain/exercise/dto/GuestInviteCommand.java
@@ -1,14 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-import umc.cockple.demo.global.enums.Gender;
-import umc.cockple.demo.global.enums.Level;
-
-@Builder
-public record GuestInviteCommand(
-        String guestName,
-        Gender gender,
-        Level level,
-        Long inviterId
-) {
-}
