--- src/main/java/umc/cockple/demo/domain/exercise/dto/GuestInviteResponseDTO.java
@@ -1,14 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record GuestInviteResponseDTO(
-        Long guestId,
-        Integer participantNumber,
-        LocalDateTime invitedAt,
-        Integer currentParticipants
-) {
-}
